verify: move Docker specific code to backends/_docker.py
