+++ Atomic/backends/_docker.py
@@ -5,6 +5,7 @@ import tempfile
 from docker import errors
 
 import Atomic.util as util
+from Atomic.mount import Mount
 from Atomic.backends.backend import Backend
 from Atomic.client import AtomicDocker, no_shaw
 from Atomic.objects.image import Image
@@ -766,4 +767,35 @@ class DockerBackend(Backend):
         if result.startswith("/"):
             result = result[1:]
         return self.d.tag(_src, result, tag=tag)
+
+    def validate_image_manifest(self):
+        """
+        Validates a docker image by mounting the image on a rootfs and validate that
+        rootfs against the manifests that were created. Note that it won't be validated
+        layer by layer.
+        :param:
+        :return: None
+        """
+        iid = self._is_image(self.image)
+        manifestname = os.path.join(util.ATOMIC_VAR_LIB, "gomtree-manifests/%s.mtree" % iid)
+        if not os.path.exists(manifestname):
+            return
+        tmpdir = tempfile.mkdtemp()
+        m = Mount()
+        m.args = []
+        m.image = self.image
+        m.mountpoint = tmpdir
+        m.mount()
+        r = util.validate_manifest(manifestname, img_rootfs=tmpdir, keywords="type,uid,gid,mode,size,sha256digest")
+        m.unmount()
+        if r.return_code != 0:
+            util.write_err(r.stdout)
+        shutil.rmtree(tmpdir)
+
+    @staticmethod
+    def get_gomtree_manifest(layer, root=os.path.join(util.ATOMIC_VAR_LIB, "gomtree-manifests")):
+        manifestpath = os.path.join(root, "%s.mtree" % layer)
+        if os.path.isfile(manifestpath):
+            return manifestpath
+        return None
